requirements.txt
Flask==2.2.1
Werkzeug==2.2.1
numpy==1.21.5
requests==2.27.1
scikit-learn==1.0.2
Pillow==9.0.1
tensorflow-cpu==2.9.1
pandas==1.4.2

